A little bit closer now. Fragments find their own ViewModel now!!!!
