+++ Cirrious/Cirrious.MvvmCross.Droid.Fragging/Fragments/MvxBindingFragmentAdapter.cs
@@ -6,30 +6,104 @@
 // Project Lead - Stuart Lodge, @slodge, me@slodge.com
 
 using System;
+using Android.OS;
+using Android.Support.V4.App;
+using Cirrious.CrossCore;
 using Cirrious.CrossCore.Core;
+using Cirrious.CrossCore.Platform;
 using Cirrious.MvvmCross.Droid.Fragging.Fragments.EventSource;
+using Cirrious.MvvmCross.Droid.Platform;
+using Cirrious.MvvmCross.ViewModels;
+using Cirrious.MvvmCross.Views;
 
 namespace Cirrious.MvvmCross.Droid.Fragging.Fragments
 {
     public class MvxBindingFragmentAdapter
         : MvxBaseFragmentAdapter
     {
-        protected IMvxFragmentView FragmentView
+        public IMvxFragmentView FragmentView
         {
-            get { return base.Fragment as IMvxFragmentView; }
+            get { return Fragment as IMvxFragmentView; }
         }
 
-        public MvxBindingFragmentAdapter(IMvxEventSourceFragment eventSource)
+        public MvxBindingFragmentAdapter(IMvxEventSourceFragment eventSource) 
             : base(eventSource)
         {
             if (!(eventSource is IMvxFragmentView))
                 throw new ArgumentException("eventSource must be an IMvxFragmentView");
         }
 
+        protected override void HandleCreateCalled(object sender, MvxValueEventArgs<Bundle> bundleArgs)
+        {
+            Bundle bundle = null;
+            MvxViewModelRequest request = null;
+            if (bundleArgs != null && bundleArgs.Value != null)
+            {
+                // saved state
+                bundle = bundleArgs.Value;
+            }
+            else
+            {
+                var fragment = FragmentView as Fragment;
+                if (fragment != null && fragment.Arguments != null)
+                {
+                    bundle = fragment.Arguments;
+                    var json = bundle.GetString("__mvxViewModelRequest");
+                    if (!string.IsNullOrEmpty(json))
+                    {
+                        IMvxNavigationSerializer serializer;
+                        if (!Mvx.TryResolve(out serializer))
+                        {
+                            MvxTrace.Warning(
+                                "Navigation Serializer not available, deserializing ViewModel Reuqest will be hard");
+                        }
+                        else
+                        {
+                            request = serializer.Serializer.DeserializeObject<MvxViewModelRequest>(json);
+                        }
+                    }
+                }    
+            }
+
+            IMvxSavedStateConverter converter;
+            if (!Mvx.TryResolve(out converter))
+            {
+                MvxTrace.Warning("Saved state converter not available - saving state will be hard");
+            }
+            else
+            {
+                if (bundle != null)
+                {
+                    var mvxBundle = converter.Read(bundle);
+                    FragmentView.OnCreate(mvxBundle, request);
+                }
+            }
+        }
+
         protected override void HandleCreateViewCalled(object sender,
-                                                       MvxValueEventArgs<MvxCreateViewParameters> args)
+                                               MvxValueEventArgs<MvxCreateViewParameters> args)
         {
             FragmentView.EnsureBindingContextIsSet(args.Value.Inflater);
+        }
+
+        protected override void HandleSaveInstanceStateCalled(object sender, MvxValueEventArgs<Bundle> bundleArgs)
+        {
+            var mvxBundle = FragmentView.CreateSaveStateBundle();
+            if (mvxBundle != null)
+            {
+                IMvxSavedStateConverter converter;
+                if (!Mvx.TryResolve(out converter))
+                {
+                    MvxTrace.Warning("Saved state converter not available - saving state will be hard");
+                }
+                else
+                {
+                    converter.Write(bundleArgs.Value, mvxBundle);
+                }
+            }
+            //TODO what to do about caching?
+            //var cache = Mvx.Resolve<IMvxSingleViewModelCache>();
+            //cache.Cache(FragmentView.ViewModel, bundleArgs.Value);
         }
 
         protected override void HandleDestroyViewCalled(object sender, EventArgs eventArgs)
